requirements.txt
flask==3.0.3
gunicorn==23.0.0
langchain-huggingface==0.0.3
python-dotenv==1.0.0
requests==2.32.3